Implemented core logic, need to fix back button and verify correct data loading.

+++ app/src/main/java/com/auto/cheapcar/entity/dto/BuildDate.kt
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName
 data class BuildDate (
         @SerializedName("wkda")
         @Expose
-        val dates: Map<Long, Long>
+        val dates: Map<Int, Int>
 )
